refactor(Settings): move landing screen to own directory
